Move misc solve Ops
